+++ src/themes.js
@@ -5,7 +5,9 @@ module.exports = {
     background2hover: 'rgb(208,208,208)',
     foreground1: 'rgb(105,105,105)',
     text1: 'rgb(36,36,36)',
-    text2: 'rgb(87,87,87)'
+    text2: 'rgb(87,87,87)',
+    // fontURL: '//fonts.googleapis.com/css?family=Roboto+Mono',
+    // fontFamily: `'Roboto Mono' monospace`,
   },
 
   dark: {
@@ -14,6 +16,8 @@ module.exports = {
     background2hover: 'rgb(58,58,58)',
     foreground1: 'rgb(112,112,112)',
     text1: 'rgb(235,235,235)',
-    text2: 'rgb(161,161,161)'
+    text2: 'rgb(161,161,161)',
+    // fontURL: '//fonts.googleapis.com/css?family=Source+Sans+Pro',
+    // fontFamily: `'Source Sans Pro', sans-serif`,
   }
 }