+++ changelog.md
@@ -1,3 +1,23 @@
+## 0.6.0
+
+- Toast notifications now have `aria-live="polite"` [accessibility]
+- Components with bound variables will now update themselves only if the bound value has changed
+    - This is still checked every frame, which is something I'd like to avoid. I'm looking into it.
+- Made component polling rate part an option in GUI `opts`
+- Text and Range components will no longer update from their bound variables while focused
+- Styling update:
+    - Component elements can now grow vertically
+    - Range now defocuses after mouseup when using the mouse (stays focused if using a keyboard) [accessibility]
+    - Checkbox now shows on/off styles when focused using a keyboard [accessibility]
+    - Button and File components now give visual feedback when clicked
+    - File component shows an outline when a file is dragged onto it
+    - Select component now highlights on mouseover or focus
+    - Fix focus highlighting issues on Firefox
+    - Add font support to themes [tentative]
+    - Changed Title component look
+- Display component text is now selectable.
+- Adjusted margin spacing for folders and titles
+
 ## 0.5.0
 
 - Add File component